reliability_preprocess function. No more numpy
Only tf is used for the reliability_preprocess function. Two main changes: use lists and then tf.stack as tf.zeros don't support assigning since tensors are immutable. Secondly, we make use of the function tf.experimental.numpy.where which substitutes the np.where function, which is used to retrieve indexes of observations that belong to the same bin.